Adding the file lambdaTrans.devR
This file is a development/ideas page for a lambda transform function.
The idea is to explore how to locally transforma t ree according to lambda and then incorporate it into a function that can locally transform a tree according to lambda, kappa, delta, OU and general branch length stretching.
There is already a function that does this for branch length stretches in geiger, but it is pretty hard to see how it works in order to include lambda, kappa and delta.
At the moment this file is just some code from phytools to explore how coding a lambda transformation works.