Bug fix in localLambda.R
Bug 1: The for loop only iterated for the second item in the list - now fixed.
Bug 2: There was a copy-paste error, I had copied a small check loop from a function with the argument scalar instead of lambda, and had not changed scalar to lambda in this file.

Both these bugs are now fixed.