added transTree.R
This is a function that will take a vector of nodes or a list of groups of tips, and apply one of either lambda, lappa, delta or rate transforms to the specified local areas of the tree. It also takes a list of parameter values to go along with each element of the nodes or tips supplied.